+++ src/ffiec_data_connect/__init__.py
@@ -12,7 +12,7 @@ shut down on February 28, 2026.
 """
 
 # Version
-__version__ = "3.0.0rc4"
+__version__ = "3.0.0rc5"
 
 # New async-compatible client
 from ffiec_data_connect.async_compatible import AsyncCompatibleClient, RateLimiter